+++ earthpy/io.py
@@ -1,11 +1,18 @@
-"""File Input/Output utilities."""
+"""
+earthpy.io
+==========
 
+File Input/Output utilities.
 
+"""
+
+import io
 import os
 import os.path as op
 import re
 import requests
-from download import download
+import tarfile
+import zipfile
 import earthpy
 
 # Data URLs, structured as {'week_name': [(URL, FILENAME, FILETYPE)]}
@@ -65,10 +72,9 @@ DATA_URLS = {
     ),
 }
 
-ALLOWED_FILE_TYPES = ["zip", "tar", "tar.gz", "file"]
-
 HOME = op.join(op.expanduser("~"))
 DATA_NAME = op.join("earth-analytics", "data")
+ALLOWED_FILE_TYPES = ["file", "tar", "tar.gz", "zip"]
 
 
 class Data(object):
@@ -110,7 +116,7 @@ class Data(object):
         s = "Available Datasets: {}".format(self.data_keys)
         return s
 
-    def get_data(self, key=None, replace=False, url=None):
+    def get_data(self, key=None, url=None, replace=False, verbose=True):
         """
         Retrieve the data for a given week and return its path.
 
@@ -129,6 +135,8 @@ class Data(object):
         replace : bool
             Whether to replace the data for this key if it is
             already downloaded.
+        verbose : bool
+            Whether to print verbose output while downloading files.
 
         Returns
         -------
@@ -165,7 +173,7 @@ class Data(object):
                 )
 
             this_data = DATA_URLS[key]
-            this_root = op.join(self.path, key)
+            this_root = op.join(str(self.path), key)
 
         if url is not None:
             with requests.head(url) as r:
@@ -175,24 +183,24 @@ class Data(object):
                 else:
                     fname = url.split("/")[-1]
 
-            # try and deduce filetype
+            # try and deduce filetype based on extension
             file_type = "file"
-            for kind in ALLOWED_FILE_TYPES:
-                if fname.endswith(kind):
-                    file_type = kind
+            for ext in ALLOWED_FILE_TYPES:
+                if fname.endswith(ext):
+                    file_type = ext
 
-            # strip off file extension so we get pretty download directories
-            if file_type != "file":
-                fname = os.path.splitext(fname)[0]
+            # remove extension for pretty download paths
+            fname = re.sub("\\.{}$".format(file_type), "", fname)
 
             this_data = (url, fname, file_type)
-            this_root = op.join(self.path, "earthpy-downloads")
+            this_root = op.join(str(self.path), "earthpy-downloads")
 
         if not isinstance(this_data, list):
             this_data = [this_data]
 
         data_paths = []
         for url, name, kind in this_data:
+
             if kind not in ALLOWED_FILE_TYPES:
                 raise ValueError(
                     "kind must be one of {}, got {}".format(
@@ -200,19 +208,96 @@ class Data(object):
                     )
                 )
 
-            # If kind is not 'file' it will be un-archived to a folder w/ `name`
-            # else create a file called `name`
-            this_path = download(
-                url,
-                os.path.join(this_root, name),
-                replace=replace,
+            this_path = self._download(
+                url=url,
+                path=os.path.join(this_root, name),
                 kind=kind,
-                verbose=False,
+                replace=replace,
+                verbose=verbose,
             )
             data_paths.append(this_path)
         if len(data_paths) == 1:
             data_paths = data_paths[0]
         return data_paths
+
+    def _download(self, url, path, kind, replace, verbose):
+        """ Download a file.
+
+        This helper function downloads files and saves them to ``path``.
+        Zip and tar files are extracted to the ``path`` directory.
+        The implementation is adapted from the download library:
+        https://github.com/choldgraf/download
+
+        Parameters
+        ----------
+        url : str
+            The URL pointing to a file to download.
+        path : str
+            Destination path of downloaded file.
+        kind: str
+            Kind of file. Must be one of ALLOWED_FILE_TYPES.
+        replace : bool
+            Whether to replace the file if it already exists.
+        verbose : bool
+            Whether to print verbose output while downloading files.
+
+
+        Returns
+        -------
+        output_path : str
+            Path to the downloaded file.
+        """
+        path = op.expanduser(path)
+        if replace is False and op.exists(path):
+            return path
+
+        if verbose is True:
+            print("Downloading from {}".format(url))
+
+        r = requests.get(url)
+
+        os.makedirs(op.dirname(path), exist_ok=True)
+        if kind == "file":
+            with open(path, "wb") as f:
+                f.write(r.content)
+        else:
+            self._download_and_extract(path, r, kind, verbose)
+        return path
+
+    def _download_and_extract(self, path, r, kind, verbose):
+        """ Download and extract a compressed archive.
+
+        This function downloads and extracts compressed directories to
+        a local directory.
+
+        Parameters
+        ----------
+        path : str
+            Destination path of downloaded file.
+        r: requests.models.Response
+            URL response that can be used to get the data.
+        kind : str
+            Kind of file. Must be one of ALLOWED_FILE_TYPES.
+        verbose : bool
+            Whether to print verbose output while downloading files.
+
+
+        Returns
+        -------
+        None
+
+        """
+        file_like_object = io.BytesIO(r.content)
+        if kind == "zip":
+            archive = zipfile.ZipFile(file_like_object)
+        if kind == "tar":
+            archive = tarfile.open(fileobj=file_like_object)
+        if kind == "tar.gz":
+            archive = tarfile.open(fileobj=file_like_object, mode="r:gz")
+        os.makedirs(path, exist_ok=True)
+        archive.extractall(path)
+        if verbose is True:
+            print("Extracted output to {}".format(path))
 
 
 def path_to_example(dataset):